Updated visit in acceptance test. Moved controller code from old modifier route to index
