+++ tests/acceptance/smoke-modifier-test.js
@@ -9,7 +9,7 @@ module('Acceptance | smoke modifier', function(hooks) {
   setupApplicationTest(hooks);
 
   test('reordering with mouse events', async function(assert) {
-    await visit('/modifier');
+    await visit('/');
 
     // when a handle is present, the element itself shall not be draggable
     assert.equal(verticalContents(), 'Uno Dos Tres Cuatro Cinco');
@@ -105,7 +105,7 @@ module('Acceptance | smoke modifier', function(hooks) {
   });
 
   test('mouse event onChange has correct dragged item', async function(assert) {
-    await visit('/modifier');
+    await visit('/');
 
     let order = findAll('[data-test-vertical-demo-handle]');
     await reorder(
@@ -119,7 +119,7 @@ module('Acceptance | smoke modifier', function(hooks) {
   });
 
   test('Test isAnimated still works without css for transitionDuration', async function(assert) {
-    await visit('/modifier');
+    await visit('/');
 
     assert.equal(verticalContents(), 'Uno Dos Tres Cuatro Cinco');
     assert.equal(horizontalContents(), 'Uno Dos Tres Cuatro Cinco');
@@ -140,7 +140,7 @@ module('Acceptance | smoke modifier', function(hooks) {
   });
 
   test('reordering with touch events', async function(assert) {
-    await visit('/modifier');
+    await visit('/');
 
     assert.equal(verticalContents(), 'Uno Dos Tres Cuatro Cinco');
     assert.equal(horizontalContents(), 'Uno Dos Tres Cuatro Cinco');
@@ -211,7 +211,7 @@ module('Acceptance | smoke modifier', function(hooks) {
   });
 
   test('Touch event onChange has correct dragged item', async function(assert) {
-    await visit('/modifier');
+    await visit('/');
 
     let order = findAll('[data-test-vertical-demo-handle]');
     await reorder(
@@ -228,7 +228,7 @@ module('Acceptance | smoke modifier', function(hooks) {
     test('A11yAudit', async function(assert) {
       assert.expect(1);
 
-      await visit("/modifier");
+      await visit("/");
       await a11yAudit();
       assert.ok(true, 'no a11y errors found!');
     });
@@ -236,7 +236,7 @@ module('Acceptance | smoke modifier', function(hooks) {
     test('Keyboard selection shows UP and DOWN visual indicators on vertical sort', async function(assert) {
       assert.expect(8);
 
-      await visit("/modifier");
+      await visit("/");
 
       const handle = find('[data-test-vertical-demo-handle]');
       await focus(handle);
@@ -266,7 +266,7 @@ module('Acceptance | smoke modifier', function(hooks) {
     test('Keyboard selection shows LEFT and RIGHT visual indicators on horizontal sort', async function(assert) {
       assert.expect(8);
 
-      await visit("/modifier");
+      await visit("/");
 
       const handle = find('[data-test-horizontal-demo-handle]');
       await focus(handle);
@@ -296,7 +296,7 @@ module('Acceptance | smoke modifier', function(hooks) {
     test('Keyboard selection is activated on ENTER', async function(assert) {
       assert.expect(3);
 
-      await visit("/modifier");
+      await visit("/");
       await focus('[data-test-vertical-demo-handle]');
       await triggerKeyEvent(
         '[data-test-vertical-demo-handle]',
@@ -316,7 +316,7 @@ module('Acceptance | smoke modifier', function(hooks) {
     test('Keyboard selection is activated on SPACE', async function(assert) {
       assert.expect(3);
 
-      await visit("/modifier");
+      await visit("/");
       await focus('[data-test-vertical-demo-handle]');
       await triggerKeyEvent(
         '[data-test-vertical-demo-handle]',
@@ -335,7 +335,7 @@ module('Acceptance | smoke modifier', function(hooks) {
 
     test('Keyboard selection is cancelled on ESC', async function(assert) {
       assert.expect(3);
-      await visit("/modifier");
+      await visit("/");
 
       await focus('[data-test-vertical-demo-handle]');
       await triggerKeyEvent(
@@ -361,7 +361,7 @@ module('Acceptance | smoke modifier', function(hooks) {
     test('Keyboard selection is cancelled on losing focus', async function(assert) {
       assert.expect(3);
 
-      await visit("/modifier");
+      await visit("/");
 
       await focus('[data-test-vertical-demo-handle]');
       await triggerKeyEvent(
@@ -384,7 +384,7 @@ module('Acceptance | smoke modifier', function(hooks) {
     test('Keyboard selection moves down on DOWN and is cancelled on ESC', async function(assert) {
       assert.expect(5);
 
-      await visit("/modifier");
+      await visit("/");
 
       await focus('[data-test-vertical-demo-handle]');
       await triggerKeyEvent(
@@ -415,7 +415,7 @@ module('Acceptance | smoke modifier', function(hooks) {
     test('Keyboard selection moves down on DOWN and is cancelled on losing focus', async function(assert) {
       assert.expect(5);
 
-      await visit("/modifier");
+      await visit("/");
 
       await focus('[data-test-vertical-demo-handle]');
       await triggerKeyEvent(
@@ -443,7 +443,7 @@ module('Acceptance | smoke modifier', function(hooks) {
     test('Keyboard selection is confirmed on ENTER', async function(assert) {
       assert.expect(5);
 
-      await visit("/modifier");
+      await visit("/");
 
       await focus('[data-test-vertical-demo-handle]');
       await triggerKeyEvent(
@@ -474,7 +474,7 @@ module('Acceptance | smoke modifier', function(hooks) {
     test('Keyboard selection moves up on UP and is confirmed on SPACE', async function(assert) {
       assert.expect(5);
 
-      await visit("/modifier");
+      await visit("/");
 
       await focus('[data-test-vertical-demo-handle]');
       await triggerKeyEvent(
@@ -515,7 +515,7 @@ module('Acceptance | smoke modifier', function(hooks) {
     test('Keyboard selection moves down on DOWN and is confirmed on SPACE', async function(assert) {
       assert.expect(5);
 
-      await visit("/modifier");
+      await visit("/");
 
       await focus('[data-test-vertical-demo-handle]');
       await triggerKeyEvent(
@@ -546,7 +546,7 @@ module('Acceptance | smoke modifier', function(hooks) {
     test('Keyboard selection moves right on RIGHT and is confirmed on ENTER', async function(assert) {
       assert.expect(5);
 
-      await visit("/modifier");
+      await visit("/");
 
       await focus('[data-test-horizontal-demo-handle]');
       await triggerKeyEvent(
@@ -577,7 +577,7 @@ module('Acceptance | smoke modifier', function(hooks) {
     test('Keyboard selection moves left on LEFT and is confirmed on ENTER', async function(assert) {
       assert.expect(5);
 
-      await visit("/modifier");
+      await visit("/");
 
       await focus('[data-test-horizontal-demo-handle]');
       await triggerKeyEvent(
@@ -616,9 +616,7 @@ module('Acceptance | smoke modifier', function(hooks) {
     });
 
     test('Keyboard event onChange has correct dragged item', async function(assert) {
-      await visit('/modifier');
-
-      await visit("/modifier");
+      await visit("/");
 
       await focus('[data-test-vertical-demo-handle]');
       await triggerKeyEvent(
