+++ tests/dummy/app/templates/index.hbs
@@ -136,7 +136,7 @@
     </section>
 
     {{#if model.dragged}}
-      <p>Just dragged {{model.dragged}}</p>
+      <p>Just dragged <span data-test-just-dragged>{{this.model.dragged}}</span></p>
     {{/if}}
   </main>
