+++ tests/dummy/app/templates/modifier.hbs
@@ -135,8 +135,8 @@
       </div>
     </section>
 
-    {{#if model.dragged}}
-      <p>Just dragged <span data-test-just-dragged>{{model.dragged}}</span></p>
+    {{#if this.model.dragged}}
+      <p>Just dragged <span data-test-just-dragged>{{this.model.dragged}}</span></p>
     {{/if}}
   </main>
